pyproject.toml
[project]
name = "smart-kyc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "boto3>=1.42.30",
    "celery>=5.6.2",
    "dataclasses>=0.8",
    "datetime>=6.0",
    "fastapi>=0.128.0",
    "google-genai>=1.59.0",
    "pydantic-settings>=2.12.0",
    "python-dateutil==2.9.0.post0",
    "redis>=7.1.0",
    "requests==2.32.3",
    "sqlalchemy==2.0.34",
    "streamlit>=1.53.0",
    "typing>=3.10.0.0",
    "uuid>=1.30",
    "uvicorn>=0.40.0",
]
